requirements.txt
# local package
-e .

# Base packages
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
sagemaker>=1.58.3
docutils<0.16,>=0.10
future>=0.15.2
pathlib>=1.0.1

# Notebooks
contextlib2>=0.6.0.post1
ipykernel>=5.2.1
ipython>=7.14.0
jupyter>=1.0.0
notebook>=6.0.3

# Analysis/Deep Learning Packages
numpy>=1.18.4
opencv-python>=4.2.0.34
pillow>=7.1.2
pycocotools>=2.0.0
scikit-image>=0.17.2
scikit-learn>=0.23.0                
scipy>=1.4.1
torch==1.2.0
torchvision==0.4.0

# Visualization
matplotlib>=3.2.1
seaborn>=0.10.1 

